requirements.txt
Flask
Jinja2
Flask-Session
Flask-SocketIO
Werkzeug
cs50
eventlet
gunicorn
python-dotenv